requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
flask-wtf==1.1.1
flask-session==0.5.0
werkzeug==2.3.7
python-dotenv==1.0.0
flask-migrate==4.0.5
email-validator==2.0.0
requests==2.31.0
pandas
numpy==1.26.4
openpyxl==3.1.2

# Keycloak Integration
authlib==1.3.0
python-keycloak==3.7.0
PyJWT==2.8.0
